+++ eazy_router/lib/src/eazy_route.dart
@@ -1,5 +1,5 @@
 import 'package:eazy_router/src/abstract_eazy_route.dart';
-import 'package:eazy_router/src/other/eazy_route_middleware.dart';
+import 'package:eazy_router/src/other/eazy_route_guard.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -8,5 +8,5 @@ abstract class EazyRoute extends AbstractEazyRoute {
   Page get page;
   bool get isAnonymous => false;
   bool get isInitial => false;
-  List<EazyRouteMiddleware> get middlewares => [];
+  List<EazyRouteGuard> get guards => [];
 }